+++ askbot/doc/source/changelog.rst
@@ -3,6 +3,8 @@ Changes in Askbot
 
 Development master branch (only on github)
 ------------------------------------------
+* Allowed localized site settings
+* Added management command `askbot_clear_moderation_queue`
 * Admins and Moderators can merge questions.
 * Improved moderation modes: flags, audit, premoderation. 
   Watched user status, IP blocking, mass content removal.